+++ cookbooks/apache2-nativex/attributes/default.rb
@@ -1,1 +1,1 @@
-default['nativex-apache2']['app_name'] = 'helloWorld'
+default['apache-nativex2']['app_name'] = 'helloWorld'