+++ cookbooks/apache2-nativex/metadata.rb
@@ -1,8 +1,8 @@
-name             'nativex-apache2'
+name             'apache-nativex2'
 maintainer       'NativeX'
 maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
-description      'Installs/Configures nativex-apache2'
+description      'Installs/Configures apache-nativex2'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'