+++ cookbooks/apache2-nativex/recipes/vhost.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-apache2
+# Cookbook Name:: apache-nativex2
 # Recipe:: vhost
 #
 # Copyright 2014, NativeX
@@ -10,7 +10,7 @@
 include_recipe "apache2"
 
 # Set up the Apache virtual host 
-web_app "#{node['nativex-apache2']['app_name']}" do 
+web_app "#{node['apache-nativex2']['app_name']}" do 
   server_name node['hostname']
   docroot "#{node['apache2']['docroot_dir']}"
   server_aliases [node['fqdn'], node['hostname']]
@@ -20,7 +20,7 @@ web_app "#{node['nativex-apache2']['app_name']}" do
   log_dir node['apache']['log_dir'] 
 end
 
-directory "#{node['apache2']['docroot_dir']}/#{node['nativex-apache2']['app_name']}" do
+directory "#{node['apache2']['docroot_dir']}/#{node['apache-nativex2']['app_name']}" do
   owner "apache" 
   group "apache"
   mode "0755"
@@ -28,7 +28,7 @@ directory "#{node['apache2']['docroot_dir']}/#{node['nativex-apache2']['app_name
   recursive true
 end
 
-template "#{node['apache2']['docroot_dir']}/#{node['nativex-apache2']['app_name']}/index.html" do
+template "#{node['apache2']['docroot_dir']}/#{node['apache-nativex2']['app_name']}/index.html" do
   source "index.html.erb"
   mode "0644"
 end